riverscapes inventory first table. More combining with rivers need space
