+++ src/reports/rpt_riverscapes_inventory/templates/template.html
@@ -109,6 +109,10 @@
           {{ figures['pie']|safe }}
           <figcaption>Segment areas grouped by feature code.</figcaption>
         </figure>
+        <figure class="report-figure">
+          {{ tables['table_of_fcodes']|safe}}
+          <figcaption>Tabular summary of segment feature codes.</figcaption>
+        </figure>
       </section>
       <section>
         <h2>Table of Stream Names and lengths</h2>
