+++ src/util/athena/__init__.py
@@ -6,9 +6,11 @@ from .athena import (
     download_file_from_s3,
     fix_s3_uri,
     get_s3_file,
+    get_metadata,
+    get_data_for_aoi,
+    run_aoi_athena_query,
+    run_athena_aoi_query
 )
-from .query_aoi import run_aoi_athena_query
-
 __all__ = [
     "athena_query_get_parsed",
     "athena_query_get_path",
