+++ src/util/pdf/__init__.py
@@ -1,0 +1,1 @@
+from .create_pdf import make_pdf_from_html
